.config/espanso/requirements.txt
jsonnet
pyyaml
